[teresa_dev RENIEC_OBTENCION adapters]

+++ src/main/java/com/pradera/praderaback/service/CategoriaService.java
@@ -92,7 +92,7 @@ public class CategoriaService {
         repositorio.save(categoria);
     }
 
-    public void eliminar(Integer id) {
+    public void eliminar(Long id) {
         repositorio.deleteById(id);
     }
 
@@ -109,7 +109,7 @@ public class CategoriaService {
         return listadto;
     }
 
-    public CategoriaDTO obtener(Integer id) {
+    public CategoriaDTO obtener(Long id) {
         CategoriaModel categoria = repositorio.findById(id).orElse(null);
         CategoriaDTO dto = null;
         if(categoria != null){
